+++ pkg/clusteragent/clusterchecks/listers_kube.go
@@ -1,0 +1,35 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-2019 Datadog, Inc.
+
+// +build clusterchecks
+// +build kubeapiserver
+
+package clusterchecks
+
+import (
+	"fmt"
+
+	"github.com/DataDog/datadog-agent/pkg/clusteragent/clusterchecks/types"
+	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/apiserver"
+)
+
+func newListers() (*types.Listers, error) {
+	ac, err := apiserver.GetAPIClient()
+	if err != nil {
+		return nil, fmt.Errorf("cannot connect to apiserver: %s", err)
+	}
+	servicesInformer := ac.InformerFactory.Core().V1().Services()
+	if servicesInformer == nil {
+		return nil, fmt.Errorf("cannot get service informer: %s", err)
+	}
+	endpointsInformer := ac.InformerFactory.Core().V1().Endpoints()
+	if endpointsInformer == nil {
+		return nil, fmt.Errorf("cannot get endpoint informer: %s", err)
+	}
+	return &types.Listers{
+		ServicesLister:  servicesInformer.Lister(),
+		EndpointsLister: endpointsInformer.Lister(),
+	}, nil
+}
